SVN_SILENT compile, --warnings
svn path=/branches/koffice/1.6/koffice/; revision=631030

+++ kexi/kexidb/connection.cpp
@@ -1374,7 +1374,7 @@ QString Connection::selectStatement( KexiDB::QuerySchema& querySchema,
 		OrderByColumnList automaticPKOrderBy;
 		const QueryColumnInfo::Vector fieldsExpanded( querySchema.fieldsExpanded() );
 		foreach (QValueVector<int>::ConstIterator, it, pkeyFieldsOrder) {
-			if ((*it) >= fieldsExpanded.count()) {
+			if ((*it) >= (int)fieldsExpanded.count()) {
 				KexiDBWarn << "Connection::selectStatement(): ORDER BY: (*it) >= fieldsExpanded.count() - " 
 					<< (*it) << " >= " << fieldsExpanded.count() << endl;
 				continue;
